hack to let the target override MBB behavior
The assumptions in MBB::getFirstTerminator() just won't hold for our target, so
we need some way to redefine it. What we do is register a class of
target-specialized functions (for now just one) with the target. The member(s)
in MachineBasicBlock then have to ask the target for a specialization. For all
other targets this change is transparent.