transitioned first instructions to new style
This is an attempt to eliminate pseudo instructions and integrate their
selection dag patterns into the flexible side-specific instructions.
The goal is to make it unneccessary to add/remove operands when going from a
pseudo (selected) instruction to a schedulable one, by having the A side
instruction of an A/B side pair match nodes in the DAG.

It might well be necessary to use actual pseudo instructions for some
operations. Loads and stores will already be tricky to handle, we'll see.